refactor(cosmx): move the reader into a cosmx/ subpackage
readers/cosmx.py + the 4 _cosmx_* helper modules -> a readers/cosmx/ package
(_reader/_io/_stitching/_discovery/_utils), with relative intra-package imports.
No behavior change; all 86 tests pass.
