+++ src/spatialdata_io/readers/cosmx/__init__.py
@@ -1,0 +1,7 @@
+"""NanoString CosMx reader."""
+
+from __future__ import annotations
+
+from ._reader import cosmx
+
+__all__ = ["cosmx"]
