+++ src/spatialdata_io/readers/cosmx/_utils.py
